ScheduledExecutionStats to spi
