+++ rundeck-data-model/src/main/java/org/rundeck/app/data/model/v1/execution/RdScheduledExecutionStats.java
@@ -1,0 +1,7 @@
+package org.rundeck.app.data.model.v1.execution;
+
+public interface RdScheduledExecutionStats {
+
+    String getContent();
+    long get_version();
+}
